requirements/requirements.txt
asgiref==3.2.3
astroid==2.3.3
colorama==0.4.3
decorator==4.4.2
Django==3.0.2
django-crispy-forms==1.8.1
django-crum==0.7.7
django-widget-tweaks==1.4.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
networkx==2.4
numpy==1.19.1
Pillow==7.0.0
pylint==2.4.4
pytz==2019.3
scikit-fuzzy==0.4.2
scipy==1.5.2
six==1.14.0
sqlparse==0.3.0
wrapt==1.11.2
